Cleanup and reorg ipc_server
