+++ plugins/ipc_server/main.cpp
@@ -1,284 +1,81 @@
-#include "aws/crt/Types.h"
-#include <cpp_api.hpp>
-#include <optional>
-#include <plugin.hpp>
-#include <util.hpp>
+#include "HeaderValue.hpp"
 
-#include <algorithm>
-#include <aws/common/array_list.h>
-#include <aws/common/byte_buf.h>
+#include <aws/common/byte_order.h>
 #include <aws/common/uuid.h>
-#include <aws/crt/Allocator.h>
 #include <aws/crt/Api.h>
-#include <aws/crt/DateTime.h>
-#include <aws/crt/JsonObject.h>
-#include <aws/crt/StlAllocator.h>
+#include <aws/crt/Types.h>
 #include <aws/crt/io/EventLoopGroup.h>
 #include <aws/crt/io/SocketOptions.h>
 #include <aws/event-stream/event_stream.h>
-#include <aws/event-stream/event_stream_rpc.h>
 #include <aws/event-stream/event_stream_rpc_server.h>
-#include <aws/http/request_response.h>
+#include <aws/io/channel_bootstrap.h>
 
-#include <chrono>
+#include <cpp_api.hpp>
+#include <plugin.hpp>
+#include <util.hpp>
+
+#include <array>
 #include <filesystem>
-#include <future>
-#include <ios>
-#include <iostream>
-#include <ostream>
-#include <stdexcept>
-#include <streambuf>
-#include <string_view>
-#include <system_error>
-#include <tuple>
-#include <type_traits>
+#include <optional>
 
 #include <cstdio>
 #include <cstring>
-#include <variant>
 
-using timestamp = std::chrono::duration<uint32_t, std::milli>;
-using bytebuffer = util::Span<uint8_t>;
-using stringbuffer = util::Span<char>;
+static void onListenerDestroy(aws_event_stream_rpc_server_listener *server, void *user_data);
 
-using HeaderValue = std::variant<
-    bool,
-    uint8_t,
-    int16_t,
-    int32_t,
-    int64_t,
-    bytebuffer,
-    stringbuffer,
-    timestamp,
-    aws_uuid>;
+//
+// Connection Management
+//
+static int onNewServerConnection(
+    aws_event_stream_rpc_server_connection *connection,
+    int error_code,
+    aws_event_stream_rpc_connection_options *connection_options,
+    void *user_data);
 
-template<typename T>
-constexpr bool is_static_value = std::is_arithmetic_v<T> || std::is_same_v<timestamp, T>
-                                 || std::is_same_v<aws_uuid, T> || std::is_same_v<T, bool>;
-template<typename T>
-constexpr bool is_dynamic_value = !is_static_value<T>;
+static void onServerConnectionShutdown(
+    aws_event_stream_rpc_server_connection *connection, int error_code, void *user_data);
 
-template<typename To, size_t N>
-// NOLINTNEXTLINE(*-c-arrays)
-To from_network_bytes(const uint8_t (&buffer)[N]) noexcept {
-    static_assert(N >= sizeof(To));
-    static_assert(std::is_trivially_default_constructible_v<To>);
-    static_assert(std::is_trivially_copy_assignable_v<To>);
+static void onProtocolMessage(
+    aws_event_stream_rpc_server_connection *connection,
+    const aws_event_stream_rpc_message_args *message_args,
+    void *user_data);
 
-    To to{};
-    if(aws_is_big_endian()) {
-        std::memcpy(&to, std::data(buffer), sizeof(To));
-        return to;
-    } else {
-        // convert from network byte order
-        auto *punned = reinterpret_cast<uint8_t *>(&to);
-        std::reverse_copy(std::begin(buffer), std::next(std::begin(buffer), sizeof(To)), punned);
-    }
-    return to;
-}
+static void sendConnectionResponse(aws_event_stream_rpc_server_connection &connection);
 
-template<size_t N, typename From>
-// NOLINTNEXTLINE(*-c-arrays)
-void to_network_bytes(uint8_t (&buffer)[N], const From &from) noexcept {
-    static_assert(N >= sizeof(From));
-    static_assert(std::is_trivially_copy_assignable_v<From>);
-    if(aws_is_big_endian()) {
-        std::memcpy(std::data(buffer), &from, sizeof(From));
-    } else {
-        auto *punned = reinterpret_cast<const uint8_t *>(&from);
-        std::reverse_copy(punned, std::next(punned, sizeof(From)), std::begin(buffer));
-    }
-}
+//
+// Stream Management
+//
+static int onIncomingStream(
+    aws_event_stream_rpc_server_connection *connection,
+    aws_event_stream_rpc_server_continuation_token *token,
+    aws_byte_cursor operation_name,
+    aws_event_stream_rpc_server_stream_continuation_options *continuation_options,
+    void *user_data);
 
-// NOLINTBEGIN(*-union-access)
-HeaderValue getValue(const aws_event_stream_header_value_pair &header) noexcept {
-    auto *staticData = std::data(header.header_value.static_val);
-    auto *variableData = header.header_value.variable_len_val;
-    switch(header.header_value_type) {
-        case AWS_EVENT_STREAM_HEADER_BOOL_TRUE:
-            return true;
-        case AWS_EVENT_STREAM_HEADER_BOOL_FALSE:
-            return false;
-        case AWS_EVENT_STREAM_HEADER_BYTE:
-            return from_network_bytes<uint8_t>(header.header_value.static_val);
-        case AWS_EVENT_STREAM_HEADER_INT16:
-            return from_network_bytes<int16_t>(header.header_value.static_val);
-        case AWS_EVENT_STREAM_HEADER_INT32:
-            return from_network_bytes<int32_t>(header.header_value.static_val);
-        case AWS_EVENT_STREAM_HEADER_INT64:
-            return from_network_bytes<int64_t>(header.header_value.static_val);
-        case AWS_EVENT_STREAM_HEADER_BYTE_BUF:
-            return bytebuffer{variableData, header.header_value_len};
-        case AWS_EVENT_STREAM_HEADER_STRING:
-            return stringbuffer{// NOLINTNEXTLINE(*-reinterpret-cast)
-                                reinterpret_cast<char *>(variableData),
-                                header.header_value_len};
-        case AWS_EVENT_STREAM_HEADER_TIMESTAMP:
-            return timestamp{from_network_bytes<uint32_t>(header.header_value.static_val)};
-        case AWS_EVENT_STREAM_HEADER_UUID: {
-            return from_network_bytes<aws_uuid>(header.header_value.static_val);
-        }
-    }
-}
+static void onContinuation(
+    aws_event_stream_rpc_server_continuation_token *token,
+    const aws_event_stream_rpc_message_args *message_args,
+    void *user_data);
 
-static inline aws_event_stream_header_value_type getType(const HeaderValue &variant) {
-    return std::visit(
-        [](auto &&value) -> aws_event_stream_header_value_type {
-            using T = std::decay_t<decltype(value)>;
-            if constexpr(std::is_same_v<bool, T>) {
-                return value ? AWS_EVENT_STREAM_HEADER_BOOL_TRUE
-                             : AWS_EVENT_STREAM_HEADER_BOOL_FALSE;
-            } else if constexpr(std::is_same_v<uint8_t, T>) {
-                return AWS_EVENT_STREAM_HEADER_BYTE;
-            } else if constexpr(std::is_same_v<int16_t, T>) {
-                return AWS_EVENT_STREAM_HEADER_INT16;
-            } else if constexpr(std::is_same_v<int32_t, T>) {
-                return AWS_EVENT_STREAM_HEADER_INT32;
-            } else if constexpr(std::is_same_v<int64_t, T>) {
-                return AWS_EVENT_STREAM_HEADER_INT64;
-            } else if constexpr(std::is_same_v<bytebuffer, T>) {
-                return AWS_EVENT_STREAM_HEADER_BYTE_BUF;
-            } else if constexpr(std::is_same_v<stringbuffer, T>) {
-                return AWS_EVENT_STREAM_HEADER_STRING;
-            } else if constexpr(std::is_same_v<timestamp, T>) {
-                return AWS_EVENT_STREAM_HEADER_TIMESTAMP;
-            } else if constexpr(std::is_same_v<aws_uuid, T>) {
-                return AWS_EVENT_STREAM_HEADER_UUID;
-            }
-        },
-        variant);
-}
+static void onContinuationClose(
+    aws_event_stream_rpc_server_continuation_token *token, void *user_data);
 
-static aws_event_stream_header_value_pair makeHeader(std::string_view name, HeaderValue var) {
-    return std::visit(
-        [&var, name](auto &&val) {
-            aws_event_stream_header_value_pair args{};
-
-            args.header_name_len =
-                name.copy(std::data(args.header_name), std::size(args.header_name));
-
-            using T = std::decay_t<decltype(val)>;
-            if constexpr(is_static_value<T>) {
-                to_network_bytes(args.header_value.static_val, val);
-                args.header_value_len = sizeof(val);
-            } else { // span or string view
-                args.header_value.variable_len_val = reinterpret_cast<uint8_t *>(std::data(val));
-                args.header_value_len = std::size(val);
-            }
-
-            args.header_value_type = getType(var);
-
-            return args;
-        },
-        var);
-}
-// NOLINTEND(*-union-access)
-
-static std::ostream &operator<<(std::ostream &os, HeaderValue v) {
-    return std::visit(
-        [&os](auto &&val) -> std::ostream & {
-            using T = std::decay_t<decltype(val)>;
-            if constexpr(std::is_arithmetic_v<T> || std::is_same_v<std::string_view, T>) {
-                os << val;
-            } else if constexpr(
-                std::is_same_v<std::true_type, T> || std::is_same_v<std::false_type, T>) {
-                auto flags = os.flags();
-                os.flags(std::ios::boolalpha);
-                os << static_cast<bool>(val);
-                os.flags(flags);
-            } else if constexpr(std::is_same_v<timestamp, T>) {
-                os << val.count() << "ms";
-            } else if constexpr(std::is_same_v<stringbuffer, T> || std::is_same_v<bytebuffer, T>) {
-                os.write(reinterpret_cast<const char *>(val.begin()), val.size());
-            } else if constexpr(std::is_same_v<T, aws_uuid>) {
-                auto flags = os.flags();
-                os.flags(std::ios::hex | std::ios::uppercase);
-                for(auto &&v : val.uuid_data) {
-                    os << v;
-                }
-                os.flags(flags);
-            }
-            return os;
-        },
-        v);
-}
-
-static auto parseHeader(const aws_event_stream_header_value_pair &pair) {
-    return std::make_pair(
-        std::string_view{std::data(pair.header_name), pair.header_name_len}, getValue(pair));
-}
-
-template<typename IntT = int64_t>
-static IntT getIntOr(const HeaderValue &var, IntT &&alternative = IntT{0}) {
-    return std::visit(
-        [&alternative](auto &&var) -> IntT {
-            using T = std::decay_t<decltype(var)>;
-            if constexpr(std::is_arithmetic_v<T>) {
-                return var;
-            } else {
-                return std::forward<IntT>(alternative);
-            }
-        },
-        var);
-}
-
-static void s_on_message_flush_fn(int error_code, void *user_data) {
-    static thread_local uintmax_t counter = 0;
-    std::cerr << "Message " << counter << " flushed: " << error_code << std::endl;
-    std::ignore = user_data;
-}
-
+//
+// Messaging
+//
 template<class SendFn, size_t N, typename DataT = std::nullptr_t>
-int sendMessage(
+static int sendMessage(
     SendFn fn,
     std::array<aws_event_stream_header_value_pair, N> &headers,
     std::optional<ggapi::Buffer> payload,
     aws_event_stream_rpc_message_type message_type,
-    uint32_t flags = 0) {
-    aws_array_list headers_list{
-        .alloc = nullptr,
-        .current_size = util::Span{headers}.size_bytes(),
-        .length = std::size(headers),
-        .item_size = sizeof(aws_event_stream_header_value_pair),
-        .data = std::data(headers),
-    };
+    uint32_t flags = 0);
 
-    Aws::Crt::Vector<uint8_t> payloadVec{};
-    if(payload.has_value()) {
-        payloadVec = payload->get<Aws::Crt::Vector<uint8_t>>(
-            0, std::min(payload->size(), uint32_t{AWS_EVENT_STREAM_MAX_MESSAGE_SIZE}));
-    }
-    auto payloadBytes = Aws::Crt::ByteBufFromArray(payloadVec.data(), payloadVec.size());
-    aws_event_stream_message message{};
-    aws_event_stream_message_init(&message, Aws::Crt::ApiAllocator(), &headers_list, &payloadBytes);
+static void onMessageFlush(int error_code, void *user_data);
 
-    struct aws_event_stream_rpc_message_args args = {
-        .headers = std::data(headers),
-        .headers_count = std::size(headers),
-        .payload = &payloadBytes,
-        .message_type = message_type,
-        .message_flags = flags,
-    };
-
-    return fn(&args);
-}
-
-static void doConnection(aws_event_stream_rpc_server_connection &connection) {
-    std::array<aws_event_stream_header_value_pair, 3> headers{
-        makeHeader(":message-type", int32_t{AWS_EVENT_STREAM_RPC_MESSAGE_TYPE_CONNECT}),
-        makeHeader(":message-flags", int32_t{0}),
-        makeHeader(":stream-id", int32_t{0})};
-
-    sendMessage(
-        [connection = &connection](auto *args) {
-            return aws_event_stream_rpc_server_connection_send_protocol_message(
-                connection, args, s_on_message_flush_fn, nullptr);
-        },
-        headers,
-        std::nullopt,
-        AWS_EVENT_STREAM_RPC_MESSAGE_TYPE_CONNECT_ACK,
-        AWS_EVENT_STREAM_RPC_MESSAGE_FLAG_CONNECTION_ACCEPTED);
-}
+//
+// Operator Overloads
+//
 
 static std::ostream &operator<<(
     std::ostream &os, const aws_event_stream_rpc_message_args &message_args) {
@@ -292,132 +89,14 @@ static std::ostream &operator<<(
         }
         os << name << '=' << value << '\n';
     }
-    std::string_view payload(
-        reinterpret_cast<const char *>(message_args.payload->buffer), message_args.payload->len);
-    return os << payload << '\n';
+    auto sv = Aws::Crt::ByteCursorToStringView(aws_byte_cursor_from_buf(message_args.payload));
+    return os.write(sv.data(), static_cast<std::streamsize>(sv.size())) << '\n';
 }
 
-static void s_fixture_on_protocol_message(
-    aws_event_stream_rpc_server_connection *connection,
-    const aws_event_stream_rpc_message_args *message_args,
-    void *user_data) {
-    std::cerr << "😹 called s_fixture_on_protocol_message" << '\n';
-    std::cerr << *message_args << '\n';
-
-    switch(message_args->message_type) {
-        case AWS_EVENT_STREAM_RPC_MESSAGE_TYPE_CONNECT:
-            doConnection(*connection);
-            return;
-        default:
-            std::cerr << "Unhandled message type " << message_args->message_type << '\n';
-            break;
-    }
-}
-
-static void on_continuation(
-    struct aws_event_stream_rpc_server_continuation_token *token,
-    const struct aws_event_stream_rpc_message_args *message_args,
-    void *user_data) {
-    std::ignore = user_data;
-    std::cerr << "🙀 called on_continutation" << std::endl;
-    std::cerr << token << '\n';
-    std::cerr << *message_args << '\n';
-
-    if(message_args->message_flags & AWS_EVENT_STREAM_RPC_MESSAGE_FLAG_TERMINATE_STREAM) {
-        std::cerr << "Stream terminating\n";
-        return;
-    }
-
-    using namespace ggapi;
-
-    auto jsonBuf = Buffer::create();
-    jsonBuf.insert(-1, util::Span{message_args->payload->buffer, message_args->payload->len});
-    auto jsonHandle = jsonBuf.fromJson();
-    auto json = jsonHandle.getHandleId() ? jsonHandle.unbox<Struct>() : Struct::create();
-
-    std::string_view name{reinterpret_cast<const char *>(user_data)};
-
-    std::cerr << "[IPC] publishToIoTCoreTopic\n";
-    auto response = Task::sendToTopic(name, json);
-    std::cerr << "[IPC] publish complete\n";
-
-    auto getStreamHeader = [message_args] {
-        for(auto &&header : util::Span{message_args->headers, message_args->headers_count}) {
-            auto &&[name, value] = parseHeader(header);
-            if(name == ":stream-id") {
-                return header;
-            }
-        }
-        return makeHeader(":stream-id", int32_t{0});
-    };
-
-    // NOLINTBEGIN(*-c-arrays)
-    // headers store their dynamic value as non-const pointer
-    // avoids const-cast, and is correct even when the dynamic value is modified by the CRT
-    char contentType[] = "application/json";
-    auto valueFromArray = [](auto &&e) { return stringbuffer{std::data(e), std::size(e) - 1}; };
-    // NOLINTEND(*-c-arrays).
-
-    auto operation = reinterpret_cast<const char *>(user_data);
-    auto oSpan = util::Span{operation, strlen(operation)}.subspan(sizeof("IPC::") - 1);
-    auto serviceModel = std::string{oSpan.data(), oSpan.size()} + "Response";
-
-    std::array headers{
-        makeHeader(":message-type", int32_t{AWS_EVENT_STREAM_RPC_MESSAGE_TYPE_APPLICATION_MESSAGE}),
-        makeHeader(":message-flags", int32_t{0}),
-        getStreamHeader(),
-        makeHeader("service-model-type", stringbuffer{serviceModel}),
-        makeHeader(":content-type", valueFromArray(contentType))};
-
-    sendMessage(
-        [token](auto *args) {
-            return aws_event_stream_rpc_server_continuation_send_message(
-                token, args, s_on_message_flush_fn, nullptr);
-        },
-        headers,
-        response.getHandleId() ? std::make_optional(response.toJson()) : std::nullopt,
-        aws_event_stream_rpc_message_type::AWS_EVENT_STREAM_RPC_MESSAGE_TYPE_APPLICATION_MESSAGE);
-}
-
-void on_continuation_close(
-    struct aws_event_stream_rpc_server_continuation_token *token, void *user_data) {
-    std::cerr << "⛔ called on_continuation_close" << std::endl;
-    std::cerr << token << '\n';
-
-    // NOLINTNEXTLINE
-    delete[] reinterpret_cast<uint8_t *>(user_data);
-}
-
-static int on_incoming_stream(
-    struct aws_event_stream_rpc_server_connection *connection,
-    struct aws_event_stream_rpc_server_continuation_token *token,
-    struct aws_byte_cursor operation_name,
-    struct aws_event_stream_rpc_server_stream_continuation_options *continuation_options,
-    void *user_data) {
-    std::cerr << "😹 called on_server_incoming_stream" << std::endl;
-
-    std::cerr << token << '\n';
-
-    std::cerr.write(
-        reinterpret_cast<const char *>(operation_name.ptr),
-        static_cast<std::ptrdiff_t>(operation_name.len))
-        << '\n';
-
-    // TODO: refactor into using a container type
-    static constexpr std::string_view prefix = "IPC::";
-    auto copy = new uint8_t[operation_name.len + prefix.size() + 1]{};
-
-    auto middle = std::copy(prefix.begin(), prefix.end(), copy);
-    std::copy_n(operation_name.ptr, operation_name.len, middle);
-
-    continuation_options->on_continuation = on_continuation;
-    continuation_options->on_continuation_closed = on_continuation_close;
-    continuation_options->user_data = copy;
-
-    return AWS_OP_SUCCESS;
-}
-
-class IpcServer : public ggapi::Plugin {
+//
+// Class Definitions
+//
+class IpcServer final : public ggapi::Plugin {
 public:
     bool onBootstrap(ggapi::Struct data) override;
     bool onBind(ggapi::Struct data) override;
@@ -429,6 +108,8 @@ public:
         static IpcServer instance{};
         return instance;
     }
+
+    std::unordered_map<aws_event_stream_rpc_server_continuation_token *, std::string> continuations;
 
 private:
     Aws::Crt::Io::EventLoopGroup eventLoop{1};
@@ -443,6 +124,7 @@ private:
     static constexpr uint16_t port = 54345;
     aws_server_bootstrap *bootstrap =
         aws_server_bootstrap_new(alloc, eventLoop.GetUnderlyingHandle());
+    aws_event_stream_rpc_server_listener *listener{};
 };
 
 // Initializes global CRT API
@@ -463,48 +145,7 @@ bool IpcServer::onBootstrap(ggapi::Struct structData) {
     return true;
 }
 
-bool IpcServer::onTerminate(ggapi::Struct structData) {
-    return true;
-}
-
-bool IpcServer::onBind(ggapi::Struct data) {
-    return true;
-}
-
-static int s_fixture_on_new_server_connection(
-    struct aws_event_stream_rpc_server_connection *connection,
-    int error_code,
-    struct aws_event_stream_rpc_connection_options *connection_options,
-    void *user_data) {
-    std::ignore = connection;
-    std::ignore = error_code;
-    std::ignore = user_data;
-
-    *connection_options = {
-        .on_incoming_stream = on_incoming_stream,
-        .on_connection_protocol_message = s_fixture_on_protocol_message,
-        .user_data = nullptr,
-    };
-
-    std::cerr << "😾 s_on_new_server_connection called " << error_code << std::endl;
-    return AWS_OP_SUCCESS;
-}
-
-static void s_fixture_on_server_connection_shutdown(
-    struct aws_event_stream_rpc_server_connection *connection, int error_code, void *user_data) {
-    std::ignore = connection;
-    std::ignore = error_code;
-    std::ignore = user_data;
-    std::cerr << "😾 s_on_server_connection_shutdown called: " << error_code << std::endl;
-}
-
-static void s_on_listener_destroy(
-    struct aws_event_stream_rpc_server_listener *server, void *user_data) {
-    std::cerr << "😾 s_on_listener_destroy called." << std::endl;
-}
-
 bool IpcServer::onStart(ggapi::Struct data) {
-    std::cerr << "onStart started 😹" << std::endl;
     static constexpr std::string_view socket_path = "/tmp/gglite-ipc.socket";
 
     if(std::filesystem::exists(socket_path)) {
@@ -516,20 +157,205 @@ bool IpcServer::onStart(ggapi::Struct data) {
         .port = port,
         .socket_options = &socketOpts.GetImpl(),
         .bootstrap = bootstrap,
-        .on_new_connection = s_fixture_on_new_server_connection,
-        .on_connection_shutdown = s_fixture_on_server_connection_shutdown,
-        .on_destroy_callback = s_on_listener_destroy,
+        .on_new_connection = onNewServerConnection,
+        .on_connection_shutdown = onServerConnectionShutdown,
+        .on_destroy_callback = onListenerDestroy,
         .user_data = nullptr,
     };
 
-    auto *listener = aws_event_stream_rpc_server_new_listener(alloc, &listenerOptions);
-
-    if(!listener) {
+    if(listener = aws_event_stream_rpc_server_new_listener(alloc, &listenerOptions); !listener) {
         int error_code = aws_last_error();
-        throw std::runtime_error("Failed to create RPC server 😿 " + std::to_string(error_code));
+        throw std::runtime_error("Failed to create RPC server: " + std::to_string(error_code));
     }
 
-    std::cerr << "Wow 🙀" << std::endl;
-
     return true;
+}
+
+bool IpcServer::onTerminate(ggapi::Struct structData) {
+    aws_event_stream_rpc_server_listener_release(listener);
+    listener = nullptr;
+    return true;
+}
+
+bool IpcServer::onBind(ggapi::Struct data) {
+    return true;
+}
+
+static void onListenerDestroy(aws_event_stream_rpc_server_listener *server, void *user_data) {
+    IpcServer::get().continuations.clear();
+}
+
+static int onNewServerConnection(
+    aws_event_stream_rpc_server_connection *connection,
+    int error_code,
+    aws_event_stream_rpc_connection_options *connection_options,
+    void *user_data) {
+    std::ignore = connection;
+    std::ignore = user_data;
+
+    *connection_options = {
+        .on_incoming_stream = onIncomingStream,
+        .on_connection_protocol_message = onProtocolMessage};
+
+    std::cerr << "[IPC] incoming connection with " << connection << '\n';
+    return AWS_OP_SUCCESS;
+}
+
+static void onServerConnectionShutdown(
+    aws_event_stream_rpc_server_connection *connection, int error_code, void *user_data) {
+    std::cerr << "[IPC] connection closed with " << connection << " with error code " << error_code
+              << '\n';
+}
+
+static void onProtocolMessage(
+    aws_event_stream_rpc_server_connection *connection,
+    const aws_event_stream_rpc_message_args *message_args,
+    void *user_data) {
+    std::cerr << "Received protocol message: " << *message_args << '\n';
+
+    switch(message_args->message_type) {
+        case AWS_EVENT_STREAM_RPC_MESSAGE_TYPE_CONNECT:
+            sendConnectionResponse(*connection);
+            return;
+        default:
+            std::cerr << "Unhandled message type " << message_args->message_type << '\n';
+            break;
+    }
+}
+
+static void sendConnectionResponse(aws_event_stream_rpc_server_connection &connection) {
+    std::array<aws_event_stream_header_value_pair, 3> headers{
+        makeHeader(":message-type", int32_t{AWS_EVENT_STREAM_RPC_MESSAGE_TYPE_CONNECT}),
+        makeHeader(":message-flags", int32_t{0}),
+        makeHeader(":stream-id", int32_t{0})};
+
+    sendMessage(
+        [connection = &connection](auto *args) {
+            return aws_event_stream_rpc_server_connection_send_protocol_message(
+                connection, args, onMessageFlush, nullptr);
+        },
+        headers,
+        std::nullopt,
+        AWS_EVENT_STREAM_RPC_MESSAGE_TYPE_CONNECT_ACK,
+        AWS_EVENT_STREAM_RPC_MESSAGE_FLAG_CONNECTION_ACCEPTED);
+}
+
+static int onIncomingStream(
+    aws_event_stream_rpc_server_connection *connection,
+    aws_event_stream_rpc_server_continuation_token *token,
+    aws_byte_cursor operation_name,
+    aws_event_stream_rpc_server_stream_continuation_options *continuation_options,
+    void *user_data) {
+    auto operation = [operation_name]() -> std::string {
+        auto sv = Aws::Crt::ByteCursorToStringView(operation_name);
+        return {sv.data(), sv.size()};
+    }();
+
+    std::cerr << "[IPC] Request for " << operation << " Received\n";
+
+    IpcServer::get().continuations.emplace(token, "IPC::" + std::move(operation));
+    *continuation_options = {
+        .on_continuation = onContinuation, .on_continuation_closed = onContinuationClose};
+
+    return AWS_OP_SUCCESS;
+}
+
+static void onContinuation(
+    aws_event_stream_rpc_server_continuation_token *token,
+    const aws_event_stream_rpc_message_args *message_args,
+    void *user_data) {
+    std::cerr << "[IPC] Continuation received:\n" << *message_args << '\n';
+
+    if(message_args->message_flags & AWS_EVENT_STREAM_RPC_MESSAGE_FLAG_TERMINATE_STREAM) {
+        std::cerr << "Stream terminating\n";
+        return;
+    }
+
+    using namespace ggapi;
+
+    auto jsonHandle =
+        Buffer::create()
+            .insert(-1, util::Span{message_args->payload->buffer, message_args->payload->len})
+            .fromJson();
+
+    auto json = jsonHandle.getHandleId() ? jsonHandle.unbox<Struct>() : Struct::create();
+
+    std::string_view name = IpcServer::get().continuations.at(token);
+
+    std::cerr << "[IPC] Publishing to LPC\n";
+    auto response = Task::sendToTopic(name, json);
+    std::cerr << "[IPC] LPC complete\n";
+
+    auto getStreamHeader = [message_args] {
+        for(auto &&header : util::Span{message_args->headers, message_args->headers_count}) {
+            auto &&[name, value] = parseHeader(header);
+            if(name == ":stream-id") {
+                return header;
+            }
+        }
+        return makeHeader(":stream-id", int32_t{0});
+    };
+
+    std::string contentType = "application/json";
+    std::string serviceModel = std::string{name.substr(sizeof("IPC::") - 1)} + "Response";
+    std::array headers{
+        makeHeader(":message-type", int32_t{AWS_EVENT_STREAM_RPC_MESSAGE_TYPE_APPLICATION_MESSAGE}),
+        makeHeader(":message-flags", int32_t{0}),
+        getStreamHeader(),
+        makeHeader("service-model-type", stringbuffer{serviceModel}),
+        makeHeader(":content-type", stringbuffer(contentType))};
+
+    sendMessage(
+        [token](auto *args) {
+            return aws_event_stream_rpc_server_continuation_send_message(
+                token, args, onMessageFlush, nullptr);
+        },
+        headers,
+        response.getHandleId() ? std::make_optional(response.toJson()) : std::nullopt,
+        aws_event_stream_rpc_message_type::AWS_EVENT_STREAM_RPC_MESSAGE_TYPE_APPLICATION_MESSAGE);
+}
+
+static void onContinuationClose(
+    aws_event_stream_rpc_server_continuation_token *token, void *user_data) {
+    IpcServer::get().continuations.erase(token);
+}
+
+template<class SendFn, size_t N, typename DataT>
+static int sendMessage(
+    SendFn fn,
+    std::array<aws_event_stream_header_value_pair, N> &headers,
+    std::optional<ggapi::Buffer> payload,
+    aws_event_stream_rpc_message_type message_type,
+    uint32_t flags) {
+    aws_array_list headers_list{
+        .alloc = nullptr,
+        .current_size = util::Span{headers}.size_bytes(),
+        .length = std::size(headers),
+        .item_size = sizeof(aws_event_stream_header_value_pair),
+        .data = std::data(headers),
+    };
+
+    Aws::Crt::Vector<uint8_t> payloadVec({'{', '}'});
+    if(payload.has_value()) {
+        payloadVec = payload->get<Aws::Crt::Vector<uint8_t>>(
+            0, std::min(payload->size(), uint32_t{AWS_EVENT_STREAM_MAX_MESSAGE_SIZE}));
+    }
+    auto payloadBytes = Aws::Crt::ByteBufFromArray(payloadVec.data(), payloadVec.size());
+    aws_event_stream_message message{};
+    aws_event_stream_message_init(&message, Aws::Crt::ApiAllocator(), &headers_list, &payloadBytes);
+
+    aws_event_stream_rpc_message_args args = {
+        .headers = std::data(headers),
+        .headers_count = std::size(headers),
+        .payload = &payloadBytes,
+        .message_type = message_type,
+        .message_flags = flags,
+    };
+
+    return fn(&args);
+}
+
+static void onMessageFlush(int error_code, void *user_data) {
+    std::ignore = user_data;
+    std::ignore = error_code;
 }
